cleanup duplicate files and test build processes
