+++ src/coherent_lasers/__init__.py
@@ -1,3 +1,1 @@
 """Coherent lasers."""
-
-__version__ = "0.1.0"
