--- src/coherent_lasers/genesis/__init__.py
@@ -1,5 +1,0 @@
-from .driver import GenesisMX
-from .mock import GenesisMXMock
-from .commands import OperationMode, Alarm, ReadWriteCmd, ReadCmd
-
-__all__ = ["GenesisMX", "GenesisMXMock", "OperationMode", "Alarm", "ReadWriteCmd", "ReadCmd"]
