+++ src/coherent_lasers/genesis_mx/__init__.py
@@ -1,31 +1,5 @@
-"""
-Driver for Coherent Genesis MX series lasers.
-
-Setup:
-- If using the HOPSDevice class, ensure the CohrHOPS DLL files are in the 'dll' directory within the 'common/hops'
-package.
-
-Classes:
-- GenesisMX: Driver for Coherent Genesis MX series lasers.
-- GenesisMXOperationModes: Enum for Genesis MX operation modes.
-- GenesisMXAlarms: Enum for Genesis MX alarms.
-- GenesisMXReadCmds: Enum for Genesis MX read commands.
-- GenesisMXWriteCmds: Enum for Genesis MX write commands.
-- GenesisMXVoxelLaser: Voxel wraper for GenesisMX class.
-"""
-
-from .commands import (
-    ReadCmds as GenesisMXReadCmds,
-    WriteCmds as GenesisMXWriteCmds,
-    OperationMode as GenesisMXOperationModes,
-    Alarm as GenesisMXAlarms,
-)
 from .driver import GenesisMX
+from .mock import GenesisMXMock
+from .commands import OperationMode, Alarm, ReadWriteCmd, ReadCmd
 
-__all__ = [
-    "GenesisMX",
-    "GenesisMXOperationModes",
-    "GenesisMXAlarms",
-    "GenesisMXReadCmds",
-    "GenesisMXWriteCmds",
-]
+__all__ = ["GenesisMX", "GenesisMXMock", "OperationMode", "Alarm", "ReadWriteCmd", "ReadCmd"]
