+++ src/coherent_lasers/genesis_mx/app/frontend/__init__.py
@@ -1,0 +1,1 @@
+"""Build files for the Web UI will be generated in ./frontend/build"""
