+++ src/coherent_lasers/genesis_mx/driver.py
@@ -1,494 +1,307 @@
-from dataclasses import dataclass
-from enum import StrEnum
+import time
 from functools import cached_property
-import logging
-import random
-from coherent_lasers.genesis_mx.commands import (
-    ReadCmds,
-    WriteCmds,
-    OperationMode,
-    Alarm,
-)
-from coherent_lasers.hops.lib import HOPSDevice, HOPSException
+
+from .hops.cohrhops import CohrHOPSDevice, HOPSCommandException
+from .base import GenesisMXInfo, LaserTemperature, LaserPower
+from .commands import Alarm, OperationMode, ReadCmd, ReadWriteCmd
 
 
-@dataclass(frozen=True)
-class GenesisMXTempMetric:
-    temp: float
-    voltage: float
+class GenesisMX(CohrHOPSDevice):
+    serial2wavelength = {"A": 488, "J": 561, "R": 639}
+    head_type2unit_factor = {"MiniX": 1000, "Mini00": 1}
+    WARMUP_TIME = 5
 
-
-@dataclass
-class GenesisMXTempMetrics:
-    main: GenesisMXTempMetric
-    etalon: GenesisMXTempMetric
-    brf: GenesisMXTempMetric
-    shg: GenesisMXTempMetric
-
-
-@dataclass(frozen=True)
-class GenesisMXHeadInfo:
-    serial: str
-    type: str
-    hours: str
-    board_revision: str
-    dio_status: str
-
-
-class GenesisMXHeadType(StrEnum):
-    """Head types for Genesis MX lasers.
-    - used in determining the unit factor for power readings and write commands.
-    """
-
-    MINIX = "MiniX"
-    MINI00 = "Mini00"
-
-
-@dataclass(frozen=True)
-class GenesisMXEnableLoop:
-    software: bool
-    interlock: bool
-    key: bool
-
-    @property
-    def enabled(self) -> bool:
-        return self.software and self.interlock and self.key
-
-    @property
-    def ready(self) -> bool:
-        return self.interlock and self.key and not self.software
-
-    def __repr__(self) -> str:
-        return f"Software: {self.software}, Interlock: {self.interlock}, Key: {self.key}"
-
-
-class GenesisMX:
-    def __init__(self, serial: str, logger: logging.Logger | None = None) -> None:
-        self.log = logger if logger else logging.getLogger(f"{__name__}.{serial}")
-        self.serial = serial
-        self.hops = HOPSDevice(self.serial)
-        if not self.hops:
-            raise ValueError(f"Failed to initialize laser: {self.serial}")
-        self.remote_control = True
-        # self.disable()
-        self._unit_factor = 1000 if self.head.type == GenesisMXHeadType.MINIX else 1
-
-    @property
-    def mode(self) -> OperationMode:
-        """Get the current mode of the laser: current mode or photomode."""
-        mode_value = self.send_read_command(ReadCmds.CURRENT_MODE)
-        return OperationMode(int(mode_value))
-
-    @mode.setter
-    def mode(self, value: OperationMode) -> None:
-        """Set the mode of the laser."""
-        self.send_write_command(WriteCmds.SET_MODE, value.value)
-
-    @property
-    def power_mw(self) -> float:
-        """Get the current power of the laser."""
-        return self.send_read_float_command(ReadCmds.POWER) * self._unit_factor
-
-    @power_mw.setter
-    def power_mw(self, value: float) -> None:
-        """Set the power of the laser."""
-        value = value / self._unit_factor
-        self.send_write_command(WriteCmds.SET_POWER, value)
-        if not self.enable_loop.enabled:
-            self.log.warning(f"Attempting to set power to {value} mW while laser is disabled.")
-
-    @property
-    def power_setpoint_mw(self) -> float:
-        """Get the current power setpoint of the laser."""
-        return float(self.send_read_command(ReadCmds.POWER_SETPOINT)) * self._unit_factor
-
-    @property
-    def ldd_current(self) -> float:
-        """
-        Get the LDD current of the laser.
-
-        Measures the current supplied to the Laser Diode Driver (LDD).
-        """
-        return float(self.send_read_command(ReadCmds.LDD_CURRENT))
-
-    @property
-    def ldd_current_limit(self) -> float:
-        """
-        Get the LDD current limit of the laser.
-
-        Measures the maximum current limit set for the Laser Diode Driver (LDD).
-        """
-        return float(self.send_read_command(ReadCmds.LDD_CURRENT_LIMIT))
-
-    @property
-    def enable_loop(self) -> GenesisMXEnableLoop:
-        return GenesisMXEnableLoop(
-            software=self.software_switch,
-            interlock=self.interlock,
-            key=self.key_switch,
-        )
-
-    @property
-    def software_switch(self) -> bool:
-        try:
-            return self.send_read_bool_command(ReadCmds.SOFTWARE_SWITCH_STATE)
-        except HOPSException:
-            return False
-
-    @property
-    def interlock(self) -> bool:
-        try:
-            return self.send_read_bool_command(ReadCmds.INTERLOCK_STATUS)
-        except HOPSException:
-            return False
-
-    @property
-    def key_switch(self) -> bool:
-        try:
-            return self.send_read_bool_command(ReadCmds.KEY_SWITCH_STATE)
-        except HOPSException:
-            return False
-
-    @property
-    def remote_control(self) -> bool:
-        try:
-            return self.send_read_bool_command(ReadCmds.REMOTE_CONTROL_STATUS)
-        except HOPSException:
-            return True
-
-    @remote_control.setter
-    def remote_control(self, value: bool) -> None:
-        try:
-            self.send_write_command(WriteCmds.SET_REMOTE_CONTROL, value)
-        except HOPSException:
-            self.log.debug(f"Failed to set remote control to {value}")
-            pass
-
-    @property
-    def analog_input(self) -> bool:
-        try:
-            return self.send_read_bool_command(ReadCmds.ANALOG_INPUT_STATUS)
-        except HOPSException:
-            return False
-
-    @analog_input.setter
-    def analog_input(self, value: bool) -> None:
-        try:
-            self.send_write_command(WriteCmds.SET_ANALOG_INPUT, value)
-        except HOPSException:
-            self.log.debug(f"Failed to set analog input to {value}")
-            pass
-
-    def enable(self) -> GenesisMXEnableLoop:
-        """Enable the laser."""
-        self.send_write_command(WriteCmds.SET_SOFTWARE_SWITCH, 1)
-        return self.enable_loop
-
-    def disable(self) -> GenesisMXEnableLoop:
-        """Disable the laser."""
-        self.send_write_command(WriteCmds.SET_SOFTWARE_SWITCH, 0)
-        return self.enable_loop
-
-    @property
-    def is_ldd_enabled(self) -> bool:
-        """
-        Get the LDD enable state of the laser.
-
-        Reads whether the Laser Diode Driver (LDD) is enabled or disabled.
-        """
-        return self.send_read_bool_command(ReadCmds.LDD_ENABLE_STATE)
-
-    # Information
+    def __init__(self, serial: str) -> None:
+        super().__init__(serial=serial)
+        self.reset()
 
     @cached_property
-    def head(self) -> GenesisMXHeadInfo:
-        """Get the laser head information."""
-        try:
-            dio_status = self.send_read_command(ReadCmds.HEAD_DIO_STATUS)
-        except HOPSException:
-            dio_status = "N/A"
-        return GenesisMXHeadInfo(
-            serial=self.send_read_command(ReadCmds.HEAD_SERIAL),
-            type=self.send_read_command(ReadCmds.HEAD_TYPE),
-            hours=self.send_read_command(ReadCmds.HEAD_HOURS),
-            board_revision=self.send_read_command(ReadCmds.HEAD_BOARD_REVISION),
-            dio_status=dio_status,
+    def info(self) -> GenesisMXInfo:
+        """Get the laser's information.
+        This includes the serial number, wavelength, head type, head hours, head DIO status, and head board revision.
+        The information is cached for the lifetime of the object.
+        :return: A GenesisMXInfo object containing the laser's information.
+        :rtype: GenesisMXInfo
+        """
+        head_type = self.send_read_command(ReadCmd.HEAD_TYPE)
+        return GenesisMXInfo(
+            serial=self.serial,
+            wavelength=self.serial2wavelength[self.serial[0]],
+            head_type=head_type,
+            head_hours=self.send_read_command(ReadCmd.HEAD_HOURS),
+            head_dio_status=None,  # Unreliable command
+            head_board_revision=self.send_read_command(ReadCmd.HEAD_BOARD_REVISION),
         )
 
-    @property
-    def alarms(self) -> list[Alarm]:
-        """Get the list of active alarms based on the fault code."""
-        fault_code_value = int(self.send_read_command(ReadCmds.FAULT_CODE), 16)
-        faults = Alarm.from_code(fault_code_value)
-        return faults
-
-    @property
-    def temperature_c(self) -> float:
+    @cached_property
+    def unit_factor(self) -> int:
+        """Unit factor for the laser's power and power setpoint.
+        Depending on the head_type, the power is returned in W or mW. We use mW as the standard unit.
         """
-        Get the main temperature of the laser.
+        if head_type := self.info.head_type:
+            return self.head_type2unit_factor.get(head_type, 1)
+        return 1
 
-        Description: Measures the temperature of the main thermoelectric cooler (TEC) that regulates
-        the overall temperature of the laser head to ensure optimal performance and stability.
+    @property
+    def power(self) -> LaserPower:
+        """Power and Power Setpoint in mW.
+        :return: The power in mW or None if an error occurred.
+        :rtype: LaserPower
         """
-        try:
-            return self.send_read_float_command(ReadCmds.MAIN_TEMPERATURE)
-        except HOPSException:
-            return 0.0
-
-    @property
-    def main_tec_drive_v(self) -> float:
-        """
-        Get the main TEC drive voltage of the laser.
-
-        Measures the drive voltage of the main thermoelectric cooler (TEC), which regulates the
-        overall temperature of the laser head.
-        """
-        try:
-            return self.send_read_float_command(ReadCmds.MAIN_TEC_DRIVE)
-        except HOPSException:
-            return 0.0
-
-    @property
-    def shg_temperature_c(self) -> float:
-        """
-        Get the SHG temperature of the laser in degrees Celsius.
-
-        Measures the temperature of the Second Harmonic Generation (SHG) heater. The SHG heater is crucial for
-        maintaining the proper temperature for efficient frequency doubling processes that convert
-        the laser light to the desired wavelength.
-        """
-        try:
-            return self.send_read_float_command(ReadCmds.SHG_TEMPERATURE)
-        except HOPSException:
-            return 0.0
-
-    @property
-    def shg_heater_drive_v(self) -> float:
-        """Get the SHG heater drive voltage of the laser."""
-        try:
-            return self.send_read_float_command(ReadCmds.SHG_HEATER_DRIVE)
-        except HOPSException:
-            return 0.0
-
-    @property
-    def brf_temperature_c(self) -> float:
-        """
-        Get the BRF temperature of the laser in degrees Celsius.
-
-        Measures the temperature of the Beam Reference Frequency (BRF) heater. The BRF heater is essential for
-        maintaining the proper temperature for the frequency reference of the laser.
-        """
-        try:
-            return self.send_read_float_command(ReadCmds.BRF_TEMPERATURE)
-        except HOPSException:
-            return 0.0
-
-    @property
-    def brf_heater_drive_v(self) -> float:
-        """Get the BRF heater drive voltage of the laser."""
-        try:
-            return self.send_read_float_command(ReadCmds.BRF_HEATER_DRIVE)
-        except HOPSException:
-            return 0.0
-
-    @property
-    def etalon_temperature_c(self) -> float:
-        """
-        Get the etalon temperature of the laser in degrees Celsius.
-
-        Measures the temperature of the etalon heater. The etalon heater is crucial for maintaining the proper
-        temperature for the etalon, which is used to stabilize the laser wavelength.
-        """
-        try:
-            return self.send_read_float_command(ReadCmds.ETALON_TEMPERATURE)
-        except HOPSException:
-            return 0.0
-
-    @property
-    def etalon_heater_drive_v(self) -> float:
-        """Get the etalon heater drive voltage of the laser."""
-        try:
-            return self.send_read_float_command(ReadCmds.ETALON_HEATER_DRIVE)
-        except HOPSException:
-            return 0.0
-
-    # Commands
-
-    def send_write_command(self, cmd: WriteCmds, new_value: float | None = None) -> None:
-        """Send a write command to the laser."""
-        self.hops.send_command(f"{cmd.value}{new_value}")
-
-    def send_read_command(self, cmd: ReadCmds) -> str:
-        """Send a read command to the laser."""
-        return self.hops.send_command(cmd.value)
-
-    def send_read_bool_command(self, cmd: ReadCmds) -> bool:
-        """Send a read command to the laser."""
-        OK = {1, "1"}
-        return self.send_read_command(cmd).strip() in OK
-
-    def send_read_float_command(self, cmd: ReadCmds) -> float:
-        """Send a read command to the laser."""
-        return float(self.send_read_command(cmd).strip())
-
-    def close(self) -> None:
-        self.power_mw = 0.0
-        self.hops.close()
-
-    def __del__(self) -> None:
-        try:
-            self.close()
-        except Exception as e:
-            self.log.debug(f"Error closing laser {self.serial}: {e}")
-
-
-class MockGenesisMX:
-    def __init__(self, serial: str, logger: logging.Logger | None = None) -> None:
-        self.log = logger if logger else logging.getLogger(f"{__name__}.{serial}")
-        self.serial = serial
-        self._power_mw = 0.0
-        self._mode = OperationMode.PHOTO
-
-        self._software_switch = False
-        self._interlock = True
-        self._key = True
-        self._remote_control_enable = True
-        self._analog_input_enable = False
-        self._head = GenesisMXHeadInfo(
-            serial=serial,
-            type=GenesisMXHeadType.MINIX,
-            hours=str(random.randint(100, 1000)),
-            board_revision="1.0",
-            dio_status="OK",
-        )
-        self._alarms = []
-
-    @property
-    def mode(self) -> OperationMode:
-        """Get the current mode of the laser: current mode or photomode."""
-        return self._mode
-
-    @mode.setter
-    def mode(self, value: OperationMode) -> None:
-        """Set the mode of the laser."""
-        self._mode = value
-
-    @property
-    def power_mw(self) -> float:
-        """Get the current power of the laser."""
-        if self.enable_loop.enabled:
-            return random.gauss(self._power_mw, 0.25)
-        return random.gauss(1.5, 0.15)
-
-    @power_mw.setter
-    def power_mw(self, value: float) -> None:
-        """Set the power of the laser."""
-        if not self.enable_loop.enabled:
-            self.log.warning(f"Attempting to set power to {value} mW while laser is disabled.")
-        self._power_mw = value
-
-    @property
-    def power_setpoint_mw(self) -> float:
-        """Get the current power setpoint of the laser."""
-        return self._power_mw
-
-    @property
-    def ldd_current(self) -> float:
-        return 8 + self.power_mw / 900
-
-    @property
-    def ldd_current_limit(self) -> float:
-        return 9 + 1000 / 900
-
-    @property
-    def enable_loop(self) -> GenesisMXEnableLoop:
-        return GenesisMXEnableLoop(
-            software=self._software_switch,
-            interlock=self._interlock,
-            key=self._key,
+        value = self.send_read_float_command(ReadCmd.POWER)
+        setpoint = self.send_read_float_command(ReadWriteCmd.POWER_SETPOINT)
+        return LaserPower(
+            value=value * self.unit_factor if value is not None else None,
+            setpoint=setpoint * self.unit_factor if setpoint is not None else None,
         )
 
-    @property
-    def software_switch(self) -> bool:
-        return self._software_switch
+    @power.setter
+    def power(self, power: float) -> None:
+        """Set the power setpoint in mW.
+        :param power: The power setpoint in mW.
+        :type power: float
+        """
+        self.send_write_command(cmd=ReadWriteCmd.POWER_SETPOINT, value=power / self.unit_factor)
 
     @property
-    def interlock(self) -> bool:
-        return self._interlock
+    def current(self) -> float | None:
+        """Current in mA.
+        :return: The current in mA or None if an error occurred.
+        :rtype: float | None
+        """
+        return self.send_read_float_command(ReadCmd.CURRENT)
 
     @property
-    def key_switch(self) -> bool:
-        return self._key
-
-    @property
-    def remote_control(self) -> bool:
-        return self._remote_control_enable
+    def remote_control(self) -> bool | None:
+        """Whether remote control is enabled.
+        :return: True if enabled, False if disabled, None if an error occurred.
+        :rtype: bool | None
+        """
+        if self.info.head_type == "MiniX":
+            return self.send_read_bool_command(ReadWriteCmd.REMOTE_CONTROL)
 
     @remote_control.setter
-    def remote_control(self, value: bool) -> None:
-        self._remote_control_enable = value
+    def remote_control(self, state: bool) -> None:
+        if self.info.head_type != "MiniX":
+            self.log.debug(f"Remote control not supported for head type: {self.info.head_type}")
+        else:
+            self.send_write_command(cmd=ReadWriteCmd.REMOTE_CONTROL, value=int(state))
 
     @property
-    def analog_input(self) -> bool:
-        return self._analog_input_enable
-
-    @analog_input.setter
-    def analog_input(self, value: bool) -> None:
-        self._analog_input_enable = value
-
-    def enable(self) -> GenesisMXEnableLoop:
-        """Enable the laser."""
-        self._software_switch = True
-        return self.enable_loop
-
-    def disable(self) -> GenesisMXEnableLoop:
-        """Disable the laser."""
-        self._software_switch = False
-        return self.enable_loop
+    def key_switch(self) -> bool | None:
+        """Key switch state.
+        :return: True if enabled, False if disabled, None if an error occurred.
+        :rtype: bool | None
+        """
+        return self.send_read_bool_command(ReadCmd.KEY_SWITCH_STATUS)
 
     @property
-    def is_ldd_enabled(self) -> bool:
-        return True
+    def interlock(self) -> bool | None:
+        """Interlock state.
+        :return: True if enabled, False if disabled, None if an error occurred.
+        :rtype: bool | None
+        """
+        return self.send_read_bool_command(ReadCmd.INTERLOCK_STATUS)
 
     @property
-    def head(self) -> GenesisMXHeadInfo:
-        return self._head
+    def software_switch(self) -> bool | None:
+        """Software switch state.
+        :return: True if enabled, False if disabled, None if an error occurred.
+        :rtype: bool | None
+        """
+        return self.send_read_bool_command(ReadWriteCmd.SOFTWARE_SWITCH)
+
+    @software_switch.setter
+    def software_switch(self, state: bool) -> None:
+        if self.interlock:  # and self.key_switch:
+            self.send_write_command(cmd=ReadWriteCmd.SOFTWARE_SWITCH, value=int(state))
+        else:
+            self.log.error(f"Cannot enable: interlock={self.interlock}, key_switch={self.key_switch}")
 
     @property
-    def alarms(self) -> list[Alarm]:
-        return self._alarms
+    def is_enabled(self) -> bool | None:
+        """Whether the laser is enabled.
+        :return: True if enabled, False if disabled, None if an error occurred.
+        :rtype: bool | None
+        """
+        return self.interlock and self.key_switch and self.software_switch
 
-    @property
-    def temperature_c(self) -> float:
-        return random.uniform(20.0, 30.0)
+    def enable(self) -> None:
+        """Enable the laser. - turns on the software switch. Requires interlock and key switch to be enabled."""
+        self.software_switch = True
+        time.sleep(self.WARMUP_TIME)
 
-    @property
-    def main_tec_drive_v(self) -> float:
-        return random.uniform(0.0, 1.0)
-
-    @property
-    def shg_temperature_c(self) -> float:
-        return random.uniform(20.0, 30.0)
-
-    @property
-    def shg_heater_drive_v(self) -> float:
-        return random.uniform(0.0, 1.0)
-
-    @property
-    def brf_temperature_c(self) -> float:
-        return random.uniform(20.0, 30.0)
-
-    @property
-    def brf_heater_drive_v(self) -> float:
-        return random.uniform(0.0, 1.0)
-
-    @property
-    def etalon_temperature_c(self) -> float:
-        return random.uniform(20.0, 30.0)
-
-    @property
-    def etalon_heater_drive_v(self) -> float:
-        return random.uniform(0.0, 1.0)
+    def disable(self) -> None:
+        """Disable the laser. - turns off the software switch."""
+        self.software_switch = False
 
     def close(self) -> None:
-        pass
+        """Disable the laser and close the connection."""
+        self.disable()
+        super().close()
+
+    def reset(self) -> None:
+        """Initialize the laser."""
+        self.remote_control = True
+        self.software_switch = False
+        self.software_switch = True
+        self.software_switch = False
+        self.power = 0
+
+    def await_power(self, max_wait_time: float = 15) -> None:
+        """Wait for the laser to reach the power setpoint."""
+        if not self.is_enabled:
+            self.log.debug("Not awaiting power. Laser is not enabled.")
+            return
+        start_time = time.time()
+        while True:
+            power = self.power
+            if power.value is not None and power.setpoint is not None and power.delta is not None:
+                allowed_delta = max(power.setpoint * 0.15, 1.5)
+                if abs(power.delta) <= allowed_delta:
+                    break
+            if time.time() - start_time > max_wait_time:
+                self.log.debug(f"Power did not reach setpoint within {max_wait_time} seconds.")
+                break
+            time.sleep(0.25)
+
+    @property
+    def analog_input(self) -> bool | None:
+        """Whether analog input control is enabled.
+        :return: True if enabled, False if disabled, None if an error occurred.
+        :rtype: bool | None
+        """
+        return self.send_read_bool_command(ReadWriteCmd.ANALOG_INPUT)
+
+    @analog_input.setter
+    def analog_input(self, state: bool) -> None:
+        self.send_write_command(cmd=ReadWriteCmd.ANALOG_INPUT, value=int(state))
+
+    @property
+    def temperature(self) -> float | None:
+        """Main temperature in °C.
+        :return: The main temperature in °C or None if an error occurred."
+        :rtype: float | None
+        """
+        return self.send_read_float_command(ReadCmd.MAIN_TEMPERATURE)
+
+    def get_temperatures(self, include_only: list[str] | None = None) -> LaserTemperature:
+        """Get the temperatures of the laser.
+
+        :param exclude: List of temperature types to exclude from the result.
+        :return: A GenesisMXTemperature object containing the temperatures.
+        """
+        temp_types = ["main", "shg", "brf", "etalon"]
+        include = include_only or temp_types
+        return LaserTemperature(
+            main=self.send_read_float_command(ReadCmd.MAIN_TEMPERATURE) if "main" in include else None,
+            shg=self.send_read_float_command(ReadCmd.SHG_TEMPERATURE) if "shg" in include else None,
+            brf=self.send_read_float_command(ReadCmd.BRF_TEMPERATURE) if "brf" in include else None,
+            etalon=None,  # Command didn't work for MiniX and Mini00
+        )
+
+    @property
+    def mode(self) -> OperationMode | None:
+        """Supports CURRENT and PHOTO modes.
+        :return: The current operation mode or None if an error occurred.
+        :rtype: OperationMode | None
+        """
+        if mode := self.send_read_int_command(ReadWriteCmd.MODE) is not None:
+            return OperationMode(mode)
+
+    @mode.setter
+    def mode(self, mode: OperationMode) -> None:
+        self.send_write_command(cmd=ReadWriteCmd.MODE, value=mode.value)
+
+    @property
+    def alarms(self) -> list[str] | None:
+        """Retrieve active alarms as a list of descriptive strings."""
+        res = self.send_read_command(ReadCmd.FAULT_CODE)
+        return Alarm.parse(int(res, 16)) if res is not None else None
+
+    def __repr__(self) -> str:
+        return f"GenesisMX(serial={self.serial}, wavelength={self.info.wavelength}, head_type={self.info.head_type})"
+
+    # send commands helper functions
+    def send_write_command(self, cmd: ReadWriteCmd, value: float | int, wait: float = 0.0) -> float | None:
+        """
+        Sends a write command, then reads back the new value from the laser.
+
+        :param cmd: The read/write command to send.
+        :param value: The value to send with the command.
+        :return: The newly-updated value from the laser if successful, None otherwise.
+        """
+        try:
+            # 1. Write the new value
+            self.send_command(command=cmd.write(value))
+
+            time.sleep(wait) if wait else None
+
+            # 2. Read back the updated value
+            if response_str := self.send_command(command=cmd.read()):
+                # 3. Attempt to parse the response as a float
+                new_value = float(response_str)
+
+                if new_value != value:
+                    self.log.debug(f"Write/readback mismatch: {value} != {new_value}")
+
+                return new_value
+
+        except (HOPSCommandException, ValueError) as e:
+            self.log.error(f"Error during write or readback: {e}")
+            return None
+
+    def send_read_command(self, cmd: ReadCmd | ReadWriteCmd) -> str | None:
+        """Send a read command to the laser.
+        :param cmd: The command to send.
+        :type cmd: GenesisMXCmd.Read
+        :return: The response from the laser or None if an error occurred.
+        :rtype: str | None
+        """
+        try:
+            return self.send_command(command=cmd.read())
+        except HOPSCommandException as e:
+            self.log.error(e)
+            return None
+
+    def send_read_bool_command(self, cmd: ReadCmd | ReadWriteCmd) -> bool | None:
+        """Send a read command to the laser and return the response as a boolean.
+        :param cmd: The command to send.
+        :type cmd: GenesisMXCmd.Read
+        :return: The response from the laser as a boolean or None if an error occurred.
+        :rtype: bool | None
+        """
+        try:
+            response = self.send_command(command=cmd.read())
+            return response == "1"
+        except HOPSCommandException as e:
+            self.log.error(e)
+            return None
+
+    def send_read_float_command(self, cmd: ReadCmd | ReadWriteCmd) -> float | None:
+        """Send a read command to the laser and return the response as a float.
+        :param cmd: The command to send.
+        :type cmd: GenesisMXCmd.Read
+        :return: The response from the laser as a float or None if an error occurred.
+        :rtype: float | None
+        """
+        try:
+            if response := self.send_command(command=cmd.read()):
+                return float(response)
+        except HOPSCommandException as e:
+            self.log.error(e)
+            return None
+
+    def send_read_int_command(self, cmd: ReadCmd | ReadWriteCmd) -> int | None:
+        """Send a read command to the laser and return the response as an int.
+        :param cmd: The command to send.
+        :type cmd: GenesisMXCmd.Read
+        :return: The response from the laser as an int or None if an error occurred.
+        :rtype: int | None
+        """
+        try:
+            if response := self.send_command(command=cmd.read()):
+                return int(response)
+        except HOPSCommandException as e:
+            self.log.error(e)
+            return None
