+++ webgui/src/styles/app.scss
@@ -7,16 +7,15 @@
 /* Global styles */
 body {
     font-family: 'Open Sans Variable', Inter, system-ui, Avenir, Helvetica, Arial, sans-serif;
-
 }
 
-// html,
-// body,
-// #app {
-//     width: 100dvw;
-//     height: 100dvh;
-//     overflow: hidden;
-// }
+html,
+body,
+#app {
+    width: 100dvw;
+    height: 100dvh;
+    overflow: hidden;
+}
 
 :root {
     line-height: 1.5;
