+++ webgui/src/styles/colors.scss
@@ -41,6 +41,8 @@
 :root {
     --cyan-400: hsl(188, 86%, 53%);
     --cyan-700: hsl(193, 82%, 31%);
+    --sky-600: hsl(200, 98%, 39%);
+    --sky-700: hsl(201, 96%, 32%);
 }
 
 :root {
@@ -55,6 +57,8 @@
     --rose-800: hsl(343, 80%, 35%);
     --rose-900: hsl(342, 75%, 30%);
     --rose-950: hsl(343, 88%, 16%);
+
+    --red-600: hsl(0, 72%, 51%);
 }
 
 :root {
